On an error during fit(), write output files (if possible)
This allows the user restart the model with the given theta. Note: depending on where the algorithm failed, the theta may be subject to another iteration of the adaptive grid - causing incompatible shapes for calculating the posterior. This too is wrapped, but potentially not exhaustibly.

Additionally, write_outputs now only relies on self.settings, instead of manually providing the arguments.